Use bootstrap keys for entire provisioning process
