+++ pkg/workflows/steps/digitalocean/utils.go
@@ -3,13 +3,7 @@ package digitalocean
 import (
 	"bytes"
 	"crypto/md5"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"encoding/pem"
 	"fmt"
-
-	"golang.org/x/crypto/ssh"
 
 	"context"
 	"encoding/base64"
@@ -38,53 +32,6 @@ func getPublicIpPort(networks []godo.NetworkV4) string {
 	}
 
 	return ""
-}
-
-func generateKeyPair(size int) (string, string, error) {
-	privateKey, err := rsa.GenerateKey(rand.Reader, size)
-
-	if err != nil {
-		return "", "", err
-	}
-
-	privateKeyPem := encodePrivateKeyToPEM(privateKey)
-	publicKey, err := generatePublicKey(&privateKey.PublicKey)
-
-	if err != nil {
-		return "", "", err
-	}
-
-	return string(privateKeyPem), string(publicKey), nil
-}
-
-// encodePrivateKeyToPEM encodes Private Key from RSA to PEM format
-func encodePrivateKeyToPEM(privateKey *rsa.PrivateKey) []byte {
-	// Get ASN.1 DER format
-	privDER := x509.MarshalPKCS1PrivateKey(privateKey)
-
-	// pem.Block
-	privBlock := pem.Block{
-		Type:    "RSA PRIVATE KEY",
-		Headers: nil,
-		Bytes:   privDER,
-	}
-
-	// Private key in PEM format
-	privatePEM := pem.EncodeToMemory(&privBlock)
-
-	return privatePEM
-}
-
-func generatePublicKey(publicKey *rsa.PublicKey) ([]byte, error) {
-	publicRsaKey, err := ssh.NewPublicKey(publicKey)
-
-	if err != nil {
-		return nil, err
-	}
-
-	pubKeyBytes := ssh.MarshalAuthorizedKey(publicRsaKey)
-
-	return pubKeyBytes, nil
 }
 
 func fingerprint(key string) (string, error) {
