+++ pkg/workflows/steps/ssh/ssh.go
@@ -27,7 +27,7 @@ func (s *Step) Run(ctx context.Context, writer io.Writer, config *steps.Config) 
 		User:    config.SshConfig.User,
 		Timeout: config.SshConfig.Timeout,
 		// TODO(stgleb): Use secure storage for private keys instead carrying them in plain text
-		Key: []byte(config.SshConfig.PrivateKey),
+		Key: []byte(config.SshConfig.BootstrapPrivateKey),
 	}
 	config.Runner, err = ssh.NewRunner(cfg)
